code refactoring for SDF.Implement collision and visual - Combine redundancy method -> MergeMeshes() method
