+++ Assets/Scripts/Tools/SDF/Util/SDF2Unity.Mesh.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 147a63151a5353426b862268bd55de26
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
